updated with main project

+++ src/Services/CompletedFormService.php
@@ -16,34 +16,46 @@ class CompletedFormService
     {
         try {
 
-            $dynamicForm = DynamicForm::findOrFail($data['dynamic_form_id']);
+            $dynamicForm = DynamicForm::find($data['dynamic_form_id']);
 
             $time = strtotime($data['expires_in']);
 
-            $completedForm = $dynamicForm->completedForms()->create([
+            $newCompletedForm = new CompletedForm([
                 'user_id' => $data['user_id'],
                 'expires_in' => date('Y-m-d', $time),
             ]);
+
+            $listNewAnswers = [];
 
             foreach ($data['answers'] as $dataAnsware) {
 
                 $field = Field::find($dataAnsware['field_id']);
 
                 if ($field->type == 'file') {
-                    $this->saveTypeFile($completedForm, $dataAnsware);
+                    $newAnsware = $this->saveTypeFile($newCompletedForm, $dataAnsware);
+
+                    $listNewAnswers[] = $newAnsware;
 
                     continue;
                 }
 
-                $completedForm->answers()->create([
+                $newAnsware = new Answare([
                     'field_id' => $dataAnsware['field_id'],
                     'answare' => json_encode($dataAnsware['answare']),
                 ]);
+
+                $listNewAnswers[] = $newAnsware;
+            }
+
+            $dynamicForm->completedForms()->save($newCompletedForm);
+
+            foreach ($listNewAnswers as $data) {
+                $newCompletedForm->answers()->save($data);
             }
         } catch (Exception $exception) {
 
-            $completedForm->answers()->delete();
-            $completedForm->delete();
+            $newCompletedForm->answers()->delete();
+            $newCompletedForm->delete();
 
             throw new HttpException(500, $exception->getMessage());
         }
@@ -60,7 +72,6 @@ class CompletedFormService
             if ($field->type == 'file') {
 
                 $this->updateTypeFile($oldAnswer, $dataAnsware);
-
                 continue;
             }
 
@@ -70,31 +81,39 @@ class CompletedFormService
         }
     }
 
-    private function saveTypeFile(CompletedForm $completedForm, array $dataAnsware): void
+    private function saveTypeFile(CompletedForm $completedForm, array $dataAnsware): Answare
     {
 
         $fileName = $this->createFileIntoStorage($dataAnsware);
 
-        $completedForm->answers()->create([
+        $newAnsware = new Answare([
             'field_id' => $dataAnsware['field_id'],
             'answare' => json_encode($fileName),
         ]);
+
+        return $newAnsware;
     }
 
     private function updateTypeFile(Answare $oldAnsware, array $newAnsware): void
     {
 
-        $file =  Storage::disk('public')->exists(json_decode($oldAnsware->answare));
+        $isBase64 = $this->isBase64($newAnsware['answare'][0]);
 
-        if ($file) {
-            Storage::disk('public')->delete(json_decode($oldAnsware->answare));
+        if($isBase64)
+        {
+            $file =  Storage::disk('public')->exists(json_decode($oldAnsware->answare));
+
+            if ($file) {
+                Storage::disk('public')->delete(json_decode($oldAnsware->answare));
+            }
+
+            $newNameFile = $this->createFileIntoStorage($newAnsware);
+    
+            $oldAnsware->answare = json_encode($newNameFile);
+    
+            $oldAnsware->save();
         }
-
-        $nameFile = $this->createFileIntoStorage($newAnsware);
-
-        $oldAnsware->answare = json_encode($nameFile);
-
-        $oldAnsware->save();
+       
     }
 
     private function createFileIntoStorage(array $dataAnsware): string
@@ -109,10 +128,19 @@ class CompletedFormService
 
         $file = str_replace(' ', '+', $file);
 
-        $fileName = uniqid() . '.' . $extension;
+        $fileName = Carbon::now()->format('d-m-Y') . '_' . uniqid() . '.' . $extension;
 
         Storage::disk('public')->put($fileName, base64_decode($file));
 
         return $fileName;
+    }
+
+    private function isBase64($content): bool
+    {
+        $first = substr($content, 0, 4);
+        if ($first == "data") {
+            return true;
+        }
+        return false;
     }
 }
